Improved Core Infrastructure: resilient Kafka, cleaner background workers, and API versioning (#61)
* Applied version to endpoints

* Fixture for domain objects

* Refactored Kafka consumer to ensure topic creation

* Updated Kiota clients with the latest API definitions

* Ensured release build to github actions

+++ src/Services/EcommerceDDD.PaymentProcessing.Tests/Application/CancelPaymentHandlerTests.cs
@@ -20,7 +20,7 @@ public class CancelPaymentHandlerTests
         var paymentWriteRepository = new DummyEventStoreRepository<Payment>();
         await paymentWriteRepository.AppendEventsAsync(payment);
 
-        var cancelPayment = CancelPayment.Create(payment.Id, PaymentCancellationReason.OrderCanceled);
+        var cancelPayment = CancelPayment.Create(payment.Id, (int)PaymentCancellationReason.OrderCanceled);
         var cancelPaymentHandler = new CancelPaymentHandler(paymentWriteRepository);
 
         // When
